requirements.txt
addict==2.4.0
click==8.0.3
jschon==0.7.3
openapi-spec-validator==0.3.1
prance==0.21.8.0
py-openapi-schema-to-json-schema==0.0.3
python-dotenv==0.19.1
PyYAML==6.0
requests==2.26.0
actions-toolkit==0.0.9
